Moved components into different folders for clarity and decoupling
