--- src/components/themes/ThemeSwitch.module.css
@@ -1,3 +1,0 @@
-.switch {
-  color: var(--font-color);
-}
